organize constants data
