+++ src/constants/logos.ts
@@ -1,0 +1,6 @@
+export const logos = {
+  design:
+    "https://res.cloudinary.com/dsxzhd4c2/image/upload/q_69/v1712986935/EZ%20Website%20-%20Design/Final_Logo_Eltezza_Design_-_White_it63xt.png",
+  creative:
+    "https://res.cloudinary.com/dsxzhd4c2/image/upload/q_69/v1712986892/EZ%20Website%20-%20Creative/Final_Logo_Eltezza_Creative_-_White_oz5nrt.png",
+};
