+++ src/constants/metadata.ts
@@ -1,0 +1,36 @@
+// METADATA
+
+export const metaData = {
+  design: {
+    title: "Eltezza | Design",
+    description:
+      "Welcome to Eltezza: home to two distinct yet intertwined identities. Our studio creates captivating narratives, while our agency arm delivers tangible results, shaping the future of brand engagement.",
+    link: "https://eltezza.com/design",
+    image:
+      "https://res.cloudinary.com/dsxzhd4c2/image/upload/v1712889804/EZ%20Website%20-%20Misc/riseabovethenoise_ffeyu4.jpg",
+  },
+  creative: {
+    title: "Eltezza | Creative",
+    description:
+      "Welcome to Eltezza: home to two distinct yet intertwined identities. Our studio creates captivating narratives, while our agency arm delivers tangible results, shaping the future of brand engagement.",
+    link: "https://eltezza.com/creative",
+    image:
+      "https://res.cloudinary.com/dsxzhd4c2/image/upload/v1712889804/EZ%20Website%20-%20Misc/riseabovethenoise_ffeyu4.jpg",
+  },
+  home: {
+    title: "Eltezza | Rise above the noise",
+    description:
+      "Welcome to Eltezza: home to two distinct yet intertwined identities. Our studio creates captivating narratives, while our agency arm delivers tangible results, shaping the future of brand engagement.",
+    link: "https://eltezza.com/",
+    image:
+      "https://res.cloudinary.com/dsxzhd4c2/image/upload/v1712889804/EZ%20Website%20-%20Misc/riseabovethenoise_ffeyu4.jpg",
+  },
+  error: {
+    title: "Page Not Found | Eltezza",
+    description:
+      "Welcome to Eltezza: home to two distinct yet intertwined identities. Our studio creates captivating narratives, while our agency arm delivers tangible results, shaping the future of brand engagement.",
+    link: "https://eltezza.com/",
+    image:
+      "https://res.cloudinary.com/dsxzhd4c2/image/upload/v1712889804/EZ%20Website%20-%20Misc/riseabovethenoise_ffeyu4.jpg",
+  },
+};
