+++ src/constants/websiteInfo.ts
@@ -1,0 +1,289 @@
+// WEBSITE INFO
+export const pageTags = {
+  design: ["Web Development", "Branding & Design", "Social Media"],
+  creative: [
+    "Promotional Videos",
+    "Advertising Imagery",
+    "Paid Ads Management",
+  ],
+};
+
+export const navigationLinks = {
+  desktop: [
+    {
+      title: "Our work",
+      href: "#work",
+    },
+    {
+      title: "What we do",
+      href: "#services",
+    },
+    {
+      title: "Clients",
+      href: "#clients",
+    },
+    {
+      title: "About us",
+      href: "#about",
+    },
+    {
+      title: "Get in touch",
+      href: "#contact",
+    },
+  ],
+
+  mobile: [
+    {
+      title: "Our work",
+      href: "#work",
+    },
+    {
+      title: "What we do",
+      href: "#services",
+    },
+    {
+      title: "Get in touch",
+      href: "#contact",
+    },
+  ],
+};
+
+export const headerDescription = {
+  design:
+    "An innovative creative powerhouse, based in a dynamic studio. We craft stories that resonate positively in our culture spanning every social channel - from user-centric web experiences to bite-sized content and beyond.",
+  creative:
+    "Enter Eltezza Creative, where we're not just content creators; we're storytellers crafting authentic, scroll-stopping content. Collaborating with amazing brands, our space thrives on genuine creativity and remarkable results.",
+};
+
+export const services = {
+  design: [
+    {
+      id: "1",
+      title: "WEB DEVELOPMENT",
+      description:
+        "From coding to launch, our web development team is your path to a digital masterpiece. Empower your business with a website that impresses and converts.",
+      wrap: true,
+    },
+    {
+      id: "2",
+      title: "WEB DESIGN",
+      description:
+        "Our web design services blend aesthetics with user experience because we understand first impressions matter, and your website is your introduction.",
+      wrap: false,
+    },
+    {
+      id: "3",
+      title: "BRANDING",
+      description:
+        "Our expert team specializes in the art of crafting distinctive and unforgettable brand identities. This includes the creation of unique logos, compelling messaging, and captivating visual elements.",
+      wrap: false,
+    },
+    {
+      id: "4",
+      title: "GRAPHIC DESIGN",
+      description:
+        "Our designs at Eltezza are not just visually pleasing; they're strategically crafted to meet your objectives. We work hand in hand with you to create attention-grabbing assets that get your message across effectively.",
+      wrap: false,
+    },
+    {
+      id: "5",
+      title: "SOCIAL MEDIA",
+      description:
+        "We start by creating a unique social media kit that aligns with your brand. Our team then curates high-quality, purposeful content strategies. We actively participate in the content collection to guarantee your social channels capture attention. With this, we then leverage our expertise and deliver captivating, scroll-stopping content.",
+      wrap: false,
+    },
+  ],
+
+  creative: [
+    {
+      id: "1",
+      title: "PROMOTIONAL VIDEOS",
+      description:
+        "We create promotional videos that are more than just stories; they're connections with your audience. From brainstorming wild ideas to bringing them to life on screen, we turn brand tales into visual experiences. Join us, and let's make your brand story one that resonates, connects, and converts.",
+      wrap: true,
+    },
+    {
+      id: "2",
+      title: "POST PRODUCTION",
+      description:
+        "Transform your content game with our post-production finesse. We take raw footage and sculpt it into polished masterpieces, ensuring each frame aligns seamlessly with your brand's identity.",
+      wrap: true,
+    },
+    {
+      id: "3",
+      title: "ADVERTISING IMAGERY",
+      description:
+        "We carefully curate striking visuals that communicate your brand's essence, creating a lasting impression across diverse channels. Releasing a new product or planning for an upcoming marketing campaign? Get in touch and let's see what we can cook up together - so many other brands already have 😉",
+      wrap: true,
+    },
+    {
+      id: "4",
+      title: "PAID ADS MANAGEMENT",
+      description:
+        "We craft ad campaigns to elevate your brand above digital noise, ensuring it resonates with your audience. From engaging ad content to meticulous tracking of performance metrics, we make sure that you are put in front of the right people. Pair that with continuous optimization, and the possibilities are endless.",
+      wrap: true,
+    },
+  ],
+};
+
+export const ourProcess = {
+  design: [
+    {
+      id: "1",
+      title: "DISCOVERY",
+    },
+    {
+      id: "2",
+      title: "AUDIT",
+    },
+    {
+      id: "3",
+      title: "CONCEPT & CREATION",
+    },
+    {
+      id: "4",
+      title: "REVIEW",
+    },
+  ],
+
+  creative: [
+    {
+      id: "1",
+      title: "DISCOVERY",
+    },
+    {
+      id: "2",
+      title: "RESEARCH",
+    },
+    {
+      id: "3",
+      title: "QUOTE",
+    },
+    {
+      id: "4",
+      title: "PRODUCTION",
+    },
+    {
+      id: "5",
+      title: "DELIVERY",
+    },
+  ],
+};
+
+export const teamDescription = {
+  design:
+    "At Eltezza, a united and passionate team works tirelessly to transform challenges into opportunities and ideas into innovation, ensuring every client's purpose is not just met but exceeded",
+  creative:
+    "Get acquainted with the team pulling the strings behind the mess and outrageous rigs...",
+};
+
+export const faqs = {
+  design: [
+    {
+      id: "1",
+      title: "How do I request a quote or estimate for a project?",
+      description:
+        "Quite simply, scroll to the bottom of our site and submit your request!",
+      wrap: false,
+    },
+    {
+      id: "2",
+      title: "What industries does your agency specialize in?",
+      description:
+        "We've worked with many different clients, ranging from construction all the way to museums. If we are a good fit, no creative endeavor is beyond our capabilities.",
+      wrap: false,
+    },
+    {
+      id: "3",
+      title: "What is the typical timeline for project completion?",
+      description:
+        "Depending on the services you require, our timelines range from 3 weeks to 12 weeks. We work on a monthly basis with a lot of our clients, let's hope we do the same for you!",
+      wrap: false,
+    },
+    {
+      id: "4",
+      title: "How much do you charge?",
+      description:
+        "We work on project-cost pricing, which ensures we can amend the work to your vision without charging for extra hours (if required). Rest assured, we don't charge for the sake of it, we carefully analyze the problem, craft the solution, and then deploy our strategies within a budget that suits both parties.",
+      wrap: false,
+    },
+    {
+      id: "5",
+      title:
+        "Do you offer ongoing support or maintenance for projects after completion?",
+      description:
+        "We offer web maintenance for our web projects, this will allow us to assist you with any updates you may have in the future on top of keeping you safe from any potential cyber-attacks. Whatever the project, we will always be here to help.",
+      wrap: false,
+    },
+    {
+      id: "6",
+      title:
+        "What sets your creative agency apart from others in the industry?",
+      description:
+        "It's quite simple, the level of care and attention we give to all of our clients is second to none. We want YOU to succeed as badly as WE want to breathe.",
+      wrap: false,
+    },
+  ],
+
+  creative: [
+    {
+      id: "1",
+      title: "How do I request a quote or estimate for a project?",
+      description:
+        "Quite simply, scroll to the bottom of our site and submit your request!",
+      wrap: false,
+    },
+    {
+      id: "2",
+      title: "What industries do you specialize in?",
+      description:
+        "Our bread & butter are FMCG brands. Spanning from drinks all the way to crackers, we have the the know-how to get you seen.",
+      wrap: false,
+    },
+    {
+      id: "3",
+      title: "What is the typical timeline for project completion?",
+      description:
+        "Depending on the services you require, our timeline ranges from 2 to 6 weeks.",
+      wrap: false,
+    },
+    {
+      id: "4",
+      title: "Does this stuff even work?",
+      description: "Well, we probably wouldn't be here if it didnt ...",
+      wrap: false,
+    },
+    {
+      id: "5",
+      title: "How much do you charge?",
+      description:
+        "We work on project-cost pricing, which ensures we can amend the work to your vision without charging for extra hours (if required). Rest assured, we don't charge for the sake of it, we carefully analyze the problem, craft the solution, and then deploy our strategies within a budget that suits both parties.",
+      wrap: false,
+    },
+    {
+      id: "6",
+      title:
+        "What sets your creative agency apart from others in the industry?",
+      description:
+        "It's quite simple, the level of care and attention we give to all of our clients is second to none. We want YOU to succeed as badly as WE want to breathe",
+      wrap: false,
+    },
+  ],
+};
+
+export const contactFormText = {
+  title: "Get in touch",
+  description:
+    "Ready to turn your ideas into digital masterpieces? Let's chat over coffee!",
+  submitted:
+    "Your connection means the world to us, and we can't wait to explore how we can make your experience even better. Whether you have questions, ideas, or just want to chat, we're here for you.",
+  error: "Something went wrong, please try again.",
+  email: "eltezza.ltd@gmail.com",
+  phone: "06 213 8994",
+};
+
+export const socialLinks = {
+  facebook: "https://www.facebook.com/eltezzaltd",
+  instagram: "https://www.instagram.com/eltezzacreative/",
+  linkedin: "https://www.linkedin.com/company/eltezza/",
+};
